LayoutTests/fast/dom/Window/area-rel-noopener-expected.txt

Test that window.opener is null when a new window is opened from an area element with rel='noopener'.


PASS: window.opener is null
